Support backend parameters via `-B name:value`
.. and expose some for lorem & chatgpt

eventually the never-used parameters about max_query_size and timeout
will be moved to backend parameters, but the main thing was to add
the ability to access gpt4:

```
$ chap ask -b openai_chatgpt -B model:gpt-4 is gpt4 smarter than gpt3.5
is gpt4 smarter than gpt3.5

As of my knowledge at this time, GPT-4 doesn't exist. Perhaps in the future,
there will be a successor to GPT-3. Usually, each newer version of a model like
this is designed to be more powerful and sophisticated, so it would likely be
"smarter" than GPT-3. However, since it doesn't currently exist, I can't make a
definite comparison.
```

:grin: